Fix #3 (Only center block in 3x3 respects harvest level)
 - Thanks to @RedBullSlurpie for pointing this out. Pretty much the only case where this was noticeable (without other mods) was Iron Hammers being able to harvest Obsidian from non-center blocks, so it would've taken quite a while to notice.

+++ src/main/java/com/alexanderstrada/practicaltools/ToolFunctions.java
@@ -52,10 +52,11 @@ public class ToolFunctions {
     public static void attemptBreak(World world, BlockPos pos, PlayerEntity player, Set<Block> effectiveOn, Set<Material> effectiveMaterials, int fortuneLevel, int silkLevel) {
         BlockState state = world.getBlockState(pos);
 
+        boolean canHarvestBlock = player.getHeldItemMainhand().canHarvestBlock(state);
         boolean isEffective = effectiveOn.contains(state.getBlock()) || effectiveMaterials.contains(state.getMaterial());
         boolean witherImmune = BlockTags.WITHER_IMMUNE.contains(state.getBlock());
 
-        if (isEffective && !witherImmune) {
+        if (canHarvestBlock && isEffective && !witherImmune) {
             world.destroyBlock(pos, false);
             Block.spawnDrops(state, world, pos, null, player, player.getHeldItemMainhand());
